Add files via upload
Sales data Analysis Project.

This project is aimed to analyze sales data to indentify trends and improve forecasting.

 

Data was downloaded from list of datasets on kaggle.com

Tools used in excel includes pivot tables and charts.

 

The analysis showed that sales was high in the month of December, with a significant sales

increase in the Northern region.

 

The data indicated that the sales rep in the Northern region made a significant turn over in 

revenue.

 

Further analysis showed that coffee was the top brand of the year. The result suggests that boosting 

marketeting efforts in other regions and products would be needed for adequate sales;

 

In conclusion, the project successfully identified key sales trends that can inform future marketing strategies.

 

I recommend allocating more resources and trainings to regions with low sales over the year and personnel 

training in order to get high sales turnaround in the coming years.